requirements.txt
django
djangorestframework
psycopg2-binary
python-decouple
pillow
djangorestframework-simplejwt
drf-yasg

dj-config-url 

django-filter

django-cors-headers

celery
redis